renamed server.ts to index.ts and added mongodb connection
