+++ server/package.json
@@ -1,20 +1,21 @@
 {
   "name": "server",
   "version": "1.0.0",
-  "main": "server.ts",
+  "main": "index.ts",
   "scripts": {
-    "dev": "nodemon --exec ts-node src/server.ts"
+    "dev": "ts-node-dev src/index.ts"
   },
   "license": "MIT",
   "dependencies": {
-    "@types/express": "^4.17.17",
     "body-parser": "^1.20.2",
     "express": "^4.18.2",
     "mongoose": "^7.0.1"
   },
   "devDependencies": {
-    "nodemon": "^2.0.21",
+    "@types/express": "^4.17.17",
+    "@types/mongoose": "^5.11.97",
     "ts-node": "^10.9.1",
+    "ts-node-dev": "^2.0.0",
     "typescript": "^4.9.5"
   }
 }
